pyproject.toml
[project]
name = "fastapi-rest-db"
version = "0.1.0"
description = "FastAPI full-stack REST demo with database"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi",
  "uvicorn",
  "python-multipart",
  "jinja2",
  "ruff>=0.15.0",
  "mysql-connector-python",
  "python-dotenv",
]

[tool.ruff]
line-length = 160
indent-width = 2

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = true
docstring-code-format = true
docstring-code-line-length = 160
